Split Fonts.kt each existing class to its own file
