+++ core/src/com/unciv/ui/components/fonts/Fonts.kt
@@ -1,257 +1,24 @@
 package com.unciv.ui.components.fonts
 
 import com.badlogic.gdx.Gdx
-import com.badlogic.gdx.graphics.Color
 import com.badlogic.gdx.graphics.GL20
 import com.badlogic.gdx.graphics.Pixmap
-import com.badlogic.gdx.graphics.Texture
 import com.badlogic.gdx.graphics.g2d.BitmapFont
-import com.badlogic.gdx.graphics.g2d.BitmapFont.BitmapFontData
-import com.badlogic.gdx.graphics.g2d.BitmapFont.Glyph
-import com.badlogic.gdx.graphics.g2d.GlyphLayout
-import com.badlogic.gdx.graphics.g2d.PixmapPacker
 import com.badlogic.gdx.graphics.g2d.SpriteBatch
 import com.badlogic.gdx.graphics.g2d.TextureRegion
 import com.badlogic.gdx.graphics.glutils.FrameBuffer
 import com.badlogic.gdx.math.Matrix4
 import com.badlogic.gdx.scenes.scene2d.Actor
 import com.badlogic.gdx.scenes.scene2d.Group
-import com.badlogic.gdx.utils.Array
-import com.badlogic.gdx.utils.Disposable
-import com.unciv.Constants
 import com.unciv.GUI
 import com.unciv.UncivGame
 import com.unciv.models.ruleset.Ruleset
-import com.unciv.models.translations.tr
 import com.unciv.ui.components.MayaCalendar
 import com.unciv.ui.components.extensions.getReadonlyPixmap
 import com.unciv.ui.components.extensions.setSize
 import com.unciv.ui.images.ImageGetter
 import com.unciv.ui.images.Portrait
 import kotlin.math.ceil
-import kotlin.math.roundToInt
-
-/** Implementations of FontImplementation will use different FontMetrics - AWT or Android.Paint,
- *  both have a class of that name, no other common point: thus we create an abstraction.
- *
- *  This is used by [Fonts.getPixmapFromActor] for vertical positioning.
- */
-class FontMetricsCommon(
-    /** (positive) distance from the baseline up to the recommended top of normal text */
-    val ascent: Float,
-    /** (positive) distance from the baseline down to the recommended bottom of normal text */
-    val descent: Float,
-    /** (positive) maximum distance from top to bottom of any text,
-     *  including potentially empty space above ascent or below descent */
-    val height: Float,
-    /** (positive) distance from the bounding box top (as defined by [height])
-     *  to the highest possible top of any text */
-
-    // Note: This is NOT what typographical leading actually is, but redefined as extra empty space
-    // on top, to make it easier to sync desktop and android. AWT has some leading but no measures
-    // outside ascent+descent+leading, while Android has its leading always 0 but typically top
-    // above ascent and bottom below descent.
-    // I chose to map AWT's spacing to the top as I found the calculations easier to visualize.
-    /** Space from the bounding box top to the top of the ascenders - includes line spacing and
-     *  room for unusually high ascenders, as [ascent] is only a recommendation. */
-    val leading: Float
-)
-
-interface FontImplementation {
-    fun setFontFamily(fontFamilyData: FontFamilyData, size: Int)
-    fun getFontSize(): Int
-    fun getCharPixmap(char: Char): Pixmap
-    fun getSystemFonts(): Sequence<FontFamilyData>
-
-    fun getBitmapFont(): BitmapFont {
-        val fontData = NativeBitmapFontData(this)
-        val font = BitmapFont(fontData, fontData.regions, false)
-        font.setOwnsTexture(true)
-        return font
-    }
-
-    fun getMetrics(): FontMetricsCommon
-}
-
-// If save in `GameSettings` need use invariantFamily.
-// If show to user need use localName.
-// If save localName in `GameSettings` may generate garbled characters by encoding.
-class FontFamilyData(
-    val localName: String,
-    val invariantName: String = localName,
-    val filePath: String? = null
-) {
-
-    @Suppress("unused") // For serialization
-    constructor() : this(default.localName, default.invariantName)
-
-    // Implement kotlin equality contract such that _only_ the invariantName field is compared.
-    override fun equals(other: Any?): Boolean {
-        return if (other is FontFamilyData) invariantName == other.invariantName
-        else super.equals(other)
-    }
-
-    override fun hashCode() = invariantName.hashCode()
-
-    /** For SelectBox usage */
-    override fun toString() = localName.tr()
-
-    companion object {
-        val default = FontFamilyData("Default Font", Fonts.DEFAULT_FONT_FAMILY)
-    }
-}
-
-// This class is loosely based on libgdx's FreeTypeBitmapFontData
-class NativeBitmapFontData(
-    private val fontImplementation: FontImplementation
-) : BitmapFontData(), Disposable {
-
-    val regions: Array<TextureRegion>
-
-    private var dirty = false
-    private val packer: PixmapPacker
-
-    private val filter = Texture.TextureFilter.Linear
-
-    private companion object {
-        /** How to get the alpha channel in a Pixmap.getPixel return value (Int) - it's the LSB */
-        const val alphaChannelMask = 255
-        /** Where to test circle for transparency */
-        // The center of a squared circle's corner wedge would be at (1-PI/4)/2 ≈ 0.1073
-        const val nearCornerRelativeOffset = 0.1f
-        /** Where to test circle for opacity */
-        // arbitrary choice just off-center
-        const val nearCenterRelativeOffset = 0.4f
-        /** Width multiplier to get extra advance after a ruleset icon, empiric */
-        const val relativeAdvanceExtra = 0.039f
-        /** Multiplier to get default kerning between a ruleset icon and 'open' characters */
-        const val relativeKerning = -0.055f
-        /** Which follower characters receive how much kerning relative to [relativeKerning] */
-        val kerningMap = mapOf('A' to 1f, 'T' to 0.6f, 'V' to 1f, 'Y' to 1.2f)
-    }
-
-    init {
-        // set general font data
-        flipped = false
-        lineHeight = fontImplementation.getFontSize().toFloat()
-        capHeight = lineHeight
-        ascent = -lineHeight
-        down = -lineHeight
-
-        // Create a packer.
-        val size = 1024
-        val packStrategy = PixmapPacker.GuillotineStrategy()
-        packer = PixmapPacker(size, size, Pixmap.Format.RGBA8888, 1, false, packStrategy)
-        packer.transparentColor = Color.WHITE
-        packer.transparentColor.a = 0f
-
-        // Generate texture regions.
-        regions = Array()
-        packer.updateTextureRegions(regions, filter, filter, false)
-
-        // Set space glyph.
-        val spaceGlyph = getGlyph(' ')
-        spaceXadvance = spaceGlyph.xadvance.toFloat()
-
-        setScale(Constants.defaultFontSize / Fonts.ORIGINAL_FONT_SIZE)
-    }
-
-    override fun getGlyph(ch: Char): Glyph = super.getGlyph(ch) ?: createAndCacheGlyph(ch)
-
-    private fun createAndCacheGlyph(ch: Char): Glyph {
-        val charPixmap = getPixmapFromChar(ch)
-
-        val glyph = Glyph()
-        glyph.id = ch.code
-        glyph.width = charPixmap.width
-        glyph.height = charPixmap.height
-        glyph.xadvance = glyph.width
-
-        // Check alpha to guess whether this is a round icon
-        // Needs to be done before disposing charPixmap, and we want to do that soon
-        val assumeRoundIcon = charPixmap.guessIsRoundSurroundedByTransparency()
-
-        val rect = packer.pack(charPixmap)
-        charPixmap.dispose()
-        glyph.page = packer.pages.size - 1 // Glyph is always packed into the last page for now.
-        glyph.srcX = rect.x.toInt()
-        glyph.srcY = rect.y.toInt()
-
-        if (ch.code >= Fonts.UNUSED_CHARACTER_CODES_START)
-            glyph.setRulesetIconGeometry(assumeRoundIcon)
-
-        // If a page was added, create a new texture region for the incrementally added glyph.
-        if (regions.size <= glyph.page)
-            packer.updateTextureRegions(regions, filter, filter, false)
-
-        setGlyphRegion(glyph, regions.get(glyph.page))
-        setGlyph(ch.code, glyph)
-        dirty = true
-
-        return glyph
-    }
-
-    private fun Pixmap.guessIsRoundSurroundedByTransparency(): Boolean {
-        // If a pixel near the center is opaque...
-        val nearCenterOffset = (width * nearCenterRelativeOffset).toInt()
-        if ((getPixel(nearCenterOffset, nearCenterOffset) and alphaChannelMask) == 0) return false
-        // ... and one near a corner is transparent ...
-        val nearCornerOffset = (width * nearCornerRelativeOffset).toInt()
-        return (getPixel(nearCornerOffset, nearCornerOffset) and alphaChannelMask) == 0
-        // ... then assume it's a circular icon surrounded by transparency - for kerning
-    }
-
-    private fun Glyph.setRulesetIconGeometry(assumeRoundIcon: Boolean) {
-        // This is a Ruleset object icon - first avoid "glue"'ing them to the next char..
-        // ends up 2px for default font scale, 1px for min, 3px for max
-        xadvance += (width * relativeAdvanceExtra).roundToInt()
-
-        if (!assumeRoundIcon) return
-
-        // Now, if we guessed it's round, do some kerning, only for the most conspicuous combos.
-        // Will look ugly for very unusual Fonts - should we limit this to only default fonts?
-
-        // Kerning is a sparse 2D array of up to 2^16 hints, each stored as byte, so this is
-        // costly: kerningMap.size * Fonts.charToRulesetImageActor.size * 512 bytes
-        // Which is 1.76MB for vanilla G&K rules.
-
-        // Ends up -3px for default font scale, -2px for minimum, -4px for max
-        val defaultKerning = (width * relativeKerning)
-        for ((char, kerning) in kerningMap)
-            setKerning(char.code, (defaultKerning * kerning).roundToInt())
-    }
-
-    private fun getPixmapForTextureName(regionName: String) =
-        Fonts.extractPixmapFromTextureRegion(ImageGetter.getDrawable(regionName).region)
-
-    private fun getPixmapFromChar(ch: Char): Pixmap {
-        return when (ch) {
-            in Fonts.allSymbols -> getPixmapForTextureName(Fonts.allSymbols[ch]!!)
-            in Fonts.charToRulesetImageActor ->
-                try {
-                    // This sometimes fails with a "Frame buffer couldn't be constructed: incomplete attachment" error, unclear why
-                    Fonts.getPixmapFromActor(Fonts.charToRulesetImageActor[ch]!!)
-                } catch (_: Exception) {
-                    Pixmap(0,0, Pixmap.Format.RGBA8888) // Empty space
-                }
-            else -> fontImplementation.getCharPixmap(ch)
-        }
-    }
-
-    override fun getGlyphs(run: GlyphLayout.GlyphRun, str: CharSequence, start: Int, end: Int, lastGlyph: Glyph?) {
-        packer.packToTexture = true // All glyphs added after this are packed directly to the texture.
-        super.getGlyphs(run, str, start, end, lastGlyph)
-        if (dirty) {
-            dirty = false
-            packer.updateTextureRegions(regions, filter, filter, false)
-        }
-    }
-
-    override fun dispose() {
-        packer.dispose()
-    }
-
-}
 
 object Fonts {
